Moved workspaces-templates
